+++ app/scripts.js
@@ -25,7 +25,7 @@ require('./styles/individu-form.css');
 require('./styles/resources.css');
 require('./styles/logement.css');
 require('./styles/redirection.css');
-require('./styles/resultat.css');
+require('./styles/resultat.scss');
 require('./styles/validation.css');
 require('./styles/text-container.css');
 require('./styles/enfants.css');
@@ -112,6 +112,7 @@ require('./js/directives/individuBlock.js');
 require('./js/directives/iframe-resizer.js');
 require('./js/directives/mailTo.js');
 require('./js/directives/nationalityChoice.js');
+require('./js/directives/offlineResult.js');
 require('./js/directives/selectOnFocus.js');
 require('./js/directives/yesNoQuestion.js');
 require('./js/directives/ym2RessourcesCallToAction.js');
