--- app/styles/resultat.css
@@ -1,34 +1,0 @@
-#feedback, #social {
-    margin-top: 1em;
-    padding-top: 1em;
-    margin-bottom: 2em;
-}
-
-#feedback ul {
-    margin-left: 0;
-    padding-left: 0;
-    list-style-type: circle;
-}
-
-#error p {
-    font-weight: bold;
-}
-
-#error small {
-    display: block;
-    margin-top: 1em;
-}
-
-#error small, #error small pre {
-    font-size: 80%;
-}
-
-#result-datetime {
-    display: none;
-}
-
-@media print {
-    #result-datetime {
-        display: initial;
-    }
-}
