+++ app/styles/resultat.scss
@@ -1,0 +1,60 @@
+@import 'commons';
+
+#feedback, #social {
+    margin-top: 1em;
+    padding-top: 1em;
+    margin-bottom: 2em;
+}
+
+#feedback ul {
+    margin-left: 0;
+    padding-left: 0;
+    list-style-type: circle;
+}
+
+#error p {
+    font-weight: bold;
+}
+
+#error small {
+    display: block;
+    margin-top: 1em;
+}
+
+#error small, #error small pre {
+    font-size: 80%;
+}
+
+#result-datetime {
+    display: none;
+}
+
+@media print {
+    #result-datetime {
+        display: initial;
+    }
+}
+
+.offline-result {
+
+    @extend .well;
+
+    .input-group  {
+        > input[type="email"] {
+            padding: 14px;
+            height: inherit;
+        }
+        .input-group-btn .btn {
+            white-space: inherit;
+            padding: 14px;
+            font-size: 14px;
+        }
+    }
+
+    &-buttons {
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
+        align-items: center;
+    }
+}
