+++ app/views/partials/foyer/layout.html
@@ -10,30 +10,6 @@
       <div ui-view="recap_situation" class="recap-situation well whiteframe-z1">
         <p class="text-center"><i class="fa fa-spinner fa-spin" aria-hidden="true"></i> Chargement en cours…</p>
       </div>
-      <form method="post" id="download" ng-hide="awaitingResults || $state.current.name !== 'foyer.resultat' || ! situation._id" class="hidden-xs">
-        <input type="hidden" name="basename" ng-value="situation._id">
-        <input type="hidden" name="base64" ng-value="base64">
-        <div class="form-group">
-          <button type="submit" class="btn btn-default btn-block" ng-click="downloadAsPdf()"
-            analytics-on="click"
-            analytics-category="Télécharger">
-            <i class="fa fa-file-pdf-o"></i>  Télécharger ma simulation
-          </button>
-        </div>
-      </form>
-      <div class="well" ng-hide="awaitingResults || $state.current.name !== 'foyer.resultat' || ! situation._id">
-        <div class="alert alert-success" ng-show="followupSubmitted">Email envoyé !</div>
-        <div class="alert alert-danger" ng-show="followupSubmitted && followupError">Une erreur s'est produite</div>
-        <form name="followup" ng-submit="submitFollowup(followup)"
-          ng-hide="awaitingResults || followupSubmitted || $state.current.name !== 'foyer.resultat' || ! situation._id">
-          <div class="form-group">
-            <input type="email" name="email" ng-model="email" placeholder="Entrez votre email" class="form-control" required />
-          </div>
-          <button type="submit" class="btn btn-default btn-block">
-            <i class="fa fa-envelope-o"></i>  Recevoir par email
-          </button>
-        </form>
-      </div>
     </aside>
   </div>
 </div>
