+++ app/views/partials/resultat.html
@@ -70,6 +70,8 @@ Erreur : {{ encodedError }}
       <p>Si vous êtes dans une situation difficile, d'<a ui-sref="sos">autres solutions existent</a>.</p>
   </div>
 
+  <offline-result situation="situation"></offline-result>
+
   <div class="print-hidden">
     <div id="feedback">
       <h4>Nous améliorons ce simulateur en continu, et vous pouvez nous y aider !</h4>
